+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added more font types to styleConstants
 - Added Android back handler
 - Added Page component
+- Added PageLoader component
 
 ### Removed